# src/api.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Books API
paths:
  /api/book/:
    get:
      description: List books
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      description: Create a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /api/book/{id}:
    get:
      description: Get a book
      parameters:
        - name: id
          in: path
          description: ID of the book
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    put:
      description: Update a book
      parameters:
        - name: id
          in: path
          description: ID of the book
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Book
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      description: Delete a book by id
      parameters:
        - name: id
          in: path
          description: ID of the book
          required: true
          schema:
            type: string
      responses:
        '204':
          description: book deleted

components:
  schemas:
    Book:
      required:
        - title
        - description
        - authors
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        authors:
          type: array
          items:
            type: string
      additionalProperties: false

    BookCreateRequest:
      properties:
        title:
          type: string
        description:
          type: string
        authors:
          type: array
          items:
            type: string
      additionalProperties: false

    BookUpdateRequest:
      properties:
        title:
          type: string
        description:
          type: string
        authors:
          type: array
          items:
            type: string
      additionalProperties: false